Camera update and removed double jump
Hey guys, I added a basic camera to our character. I used the cinemachine package, which you guys will probably have to download on your own. How it works is basically one camera object in unity will have the CineMachine Brain component, then you can create multiple CineMachine objects which are like very customizable cameras. The camera with the CineMachine Brain will switch to whichever cinemachine camera has the highest priority, which is a number we control. In this case, under the Demo_2 scene, main camera has the brain, and player camera is the cinemachine that follows the player. I tweaked the camera so that it moves a little bit in front of the character when he's moving, that way the player can see a little more ahead. If you have any thoughts or questions lmk. Anyways this message is getting pretty long so I'll stop it here.